Remove commented-out legacy functions
There are a lot of functions in the Python 2 version of HostDoc (such as the ones being deleted here) that may or may not be used.  Instead of porting them all over to cyhy-db, we plan to port them as they are needed and rewrite them to take advantage of Python 3.